third_party/WebKit/LayoutTests/platform/mac-snowleopard/virtual/web-animations-css/animations/transition-and-animation-3-expected.txt

Warning this test is running in real-time and may be flaky.
FAIL - "left" property for "test" element at 0.2s expected: 100 but saw: 500.9979248046875

Log:
11.685999999940577 - runAnimationTest
13.231000000018867 - First webkitAnimationStart event fired
13.272000000029038 - Finished waiting for animations to start
13.323000000013963 - RAF callback, animation time: 0.03299999991668301
13.345999999955893 - Running checks for time: 0, delay: 0.03299999991668301
13.508999999999105 - RAF callback, animation time: 0.2250000000003638
14.293999999949847 - RAF callback, animation time: 1.0089999999536303
15.412999999966814 - RAF callback, animation time: 2.127999999970598
16.53899999996611 - RAF callback, animation time: 3.253999999969892
17.643000000020947 - RAF callback, animation time: 4.357999999911044
18.76900000002024 - RAF callback, animation time: 5.483999999910338
19.887999999923522 - RAF callback, animation time: 6.602999999927306
20.966999999927793 - RAF callback, animation time: 7.682999999929052
22.082999999952335 - RAF callback, animation time: 8.798999999953594
23.198999999976877 - RAF callback, animation time: 9.914999999978136
24.332999999955973 - RAF callback, animation time: 11.04299999997238
25.42099999993752 - RAF callback, animation time: 12.13299999994888
26.54599999993934 - RAF callback, animation time: 13.260999999943124
27.644000000009328 - RAF callback, animation time: 14.359000000013111
28.76200000002882 - RAF callback, animation time: 15.476999999918917
29.871999999954824 - RAF callback, animation time: 16.586999999958607
30.98899999997684 - RAF callback, animation time: 17.7049999999781
32.077999999955864 - RAF callback, animation time: 18.793999999957123
33.17600000002585 - RAF callback, animation time: 19.890999999915948
34.302000000025146 - RAF callback, animation time: 21.016999999915242
35.412999999948624 - RAF callback, animation time: 22.128999999949883
36.52999999997064 - RAF callback, animation time: 23.244999999974425
37.647999999990134 - RAF callback, animation time: 24.362999999993917
38.745999999946434 - RAF callback, animation time: 25.460999999950218
39.878999999928055 - RAF callback, animation time: 26.59399999993184
40.99799999994502 - RAF callback, animation time: 27.712999999948806
42.12399999994432 - RAF callback, animation time: 28.8389999999481
43.24399999995876 - RAF callback, animation time: 29.958999999962543
44.464999999945576 - RAF callback, animation time: 31.17999999994936
45.604000000025735 - RAF callback, animation time: 32.31799999991836
46.71599999994669 - RAF callback, animation time: 33.43099999995047
47.83599999996113 - RAF callback, animation time: 34.550999999964915
49.048999999968146 - RAF callback, animation time: 35.76399999997193
50.21999999996751 - RAF callback, animation time: 36.930999999981395
51.360999999928936 - RAF callback, animation time: 38.07299999994029
52.462999999988824 - RAF callback, animation time: 39.17799999999261
53.67400000000089 - RAF callback, animation time: 40.38900000000467
54.830999999921914 - RAF callback, animation time: 41.54399999993075
55.97099999999955 - RAF callback, animation time: 42.68700000000081
57.1189999999433 - RAF callback, animation time: 43.830999999954656
58.26200000001336 - RAF callback, animation time: 44.97700000001714
59.4380000000001 - RAF callback, animation time: 46.151000000008935
60.59099999993123 - RAF callback, animation time: 47.30399999994006
61.73400000000129 - RAF callback, animation time: 48.44700000001012
62.87499999996271 - RAF callback, animation time: 49.59099999996397
64.0449999999646 - RAF callback, animation time: 50.74999999999363
65.18099999993865 - RAF callback, animation time: 51.89699999993991
66.32799999999861 - RAF callback, animation time: 53.04100000000744
67.47099999995498 - RAF callback, animation time: 54.18599999995877
68.62499999999727 - RAF callback, animation time: 55.338000000006105
69.74800000000414 - RAF callback, animation time: 56.46300000000792
70.90699999992012 - RAF callback, animation time: 57.61999999992895
72.03199999992194 - RAF callback, animation time: 58.747999999923195
73.19399999994403 - RAF callback, animation time: 59.90699999995286
74.312999999961 - RAF callback, animation time: 61.02799999996478
75.44399999994766 - RAF callback, animation time: 62.152999999966596
76.57299999993938 - RAF callback, animation time: 63.28699999994569
77.70099999993363 - RAF callback, animation time: 64.38500000001568
78.83099999992282 - RAF callback, animation time: 65.54599999992661
79.97299999999541 - RAF callback, animation time: 66.68600000000424
81.0950000000048 - RAF callback, animation time: 67.81000000000859
82.22599999999147 - RAF callback, animation time: 68.94199999999273
83.34500000000844 - RAF callback, animation time: 70.05800000001727
84.46600000002036 - RAF callback, animation time: 71.17699999992055
85.59200000001965 - RAF callback, animation time: 72.30399999991732
86.74899999994068 - RAF callback, animation time: 73.45999999995456
87.92399999992995 - RAF callback, animation time: 74.6359999999413
89.0550000000303 - RAF callback, animation time: 75.7699999999204
90.18400000002202 - RAF callback, animation time: 76.89899999991212
91.36200000000372 - RAF callback, animation time: 78.0770000000075
92.47900000002574 - RAF callback, animation time: 79.19299999991836
93.56300000001738 - RAF callback, animation time: 80.27799999990748
94.65899999997873 - RAF callback, animation time: 81.37499999997999
95.7419999999729 - RAF callback, animation time: 82.45699999997669
96.8359999999393 - RAF callback, animation time: 83.53999999997086
97.91499999994357 - RAF callback, animation time: 84.62999999994736
98.9999999999327 - RAF callback, animation time: 85.71499999993648
100.08800000002793 - RAF callback, animation time: 86.80299999991803
101.17400000001453 - RAF callback, animation time: 87.88899999990463
102.27099999997336 - RAF callback, animation time: 88.98599999997714
103.38500000000295 - RAF callback, animation time: 90.10000000000673
104.50300000002244 - RAF callback, animation time: 91.21799999991254
105.61699999993834 - RAF callback, animation time: 92.33199999994213
106.74499999993259 - RAF callback, animation time: 93.45999999993637
107.8619999999546 - RAF callback, animation time: 94.57699999995839
108.98599999995895 - RAF callback, animation time: 95.70099999996273
110.09400000000369 - RAF callback, animation time: 96.80900000000747
111.24999999992724 - RAF callback, animation time: 97.96499999993102
112.40999999995438 - RAF callback, animation time: 99.12499999995816
113.58499999994365 - RAF callback, animation time: 100.29499999996005
114.73799999998846 - RAF callback, animation time: 101.45299999999224
115.88999999992211 - RAF callback, animation time: 102.6049999999259
117.04899999995177 - RAF callback, animation time: 103.76399999995556
118.20399999999154 - RAF callback, animation time: 104.91899999999532
119.35299999993276 - RAF callback, animation time: 106.06899999993402
120.5049999999801 - RAF callback, animation time: 107.22099999998136
121.69199999993907 - RAF callback, animation time: 108.40699999994285
122.81499999994594 - RAF callback, animation time: 109.5309999999472
123.96899999998823 - RAF callback, animation time: 110.68399999999201
125.12099999992188 - RAF callback, animation time: 111.83599999992566
126.24399999992875 - RAF callback, animation time: 112.95899999993253
127.40199999996094 - RAF callback, animation time: 114.11699999996472
128.52299999997285 - RAF callback, animation time: 115.23899999997411
129.6909999999798 - RAF callback, animation time: 116.40599999998358
130.81999999997151 - RAF callback, animation time: 117.53599999997277
131.97300000001633 - RAF callback, animation time: 118.68799999990642
133.12099999996008 - RAF callback, animation time: 119.83699999996134
134.27500000000236 - RAF callback, animation time: 120.99000000000615
135.42499999994106 - RAF callback, animation time: 122.13999999994485
136.57999999998083 - RAF callback, animation time: 123.29599999998209
137.73600000001807 - RAF callback, animation time: 124.44499999992331
138.89099999994414 - RAF callback, animation time: 125.60199999995803
140.04099999999653 - RAF callback, animation time: 126.75699999999779
141.18899999994028 - RAF callback, animation time: 127.90299999994659
142.34099999998762 - RAF callback, animation time: 129.05699999998888
143.50899999999456 - RAF callback, animation time: 130.22399999999834
144.65999999993073 - RAF callback, animation time: 131.37499999993452
145.81199999997807 - RAF callback, animation time: 132.52699999998185
146.9640000000254 - RAF callback, animation time: 133.67999999991298
148.11199999996916 - RAF callback, animation time: 134.82799999997042
149.24599999994825 - RAF callback, animation time: 135.96099999995204
150.4410000000007 - RAF callback, animation time: 137.15400000000955
151.6149999999925 - RAF callback, animation time: 138.3299999999963
152.75599999995393 - RAF callback, animation time: 139.4709999999577
153.92499999995835 - RAF callback, animation time: 140.63999999996213
155.0680000000284 - RAF callback, animation time: 141.7829999999185
156.2339999999267 - RAF callback, animation time: 142.94999999992797
157.39899999994122 - RAF callback, animation time: 144.113999999945
158.56399999995574 - RAF callback, animation time: 145.27899999995952
159.73699999995006 - RAF callback, animation time: 146.45199999995384
160.90399999995952 - RAF callback, animation time: 147.61999999996078
162.07199999996647 - RAF callback, animation time: 148.78699999997025
163.23799999997846 - RAF callback, animation time: 149.95399999997971
164.40699999998287 - RAF callback, animation time: 151.12299999998413
165.57799999998224 - RAF callback, animation time: 152.2939999999835
166.74699999998666 - RAF callback, animation time: 153.46299999998791
167.9229999999734 - RAF callback, animation time: 154.63799999997718
169.09599999996772 - RAF callback, animation time: 155.81199999996898
170.2339999999367 - RAF callback, animation time: 156.94999999993797
171.36900000002697 - RAF callback, animation time: 158.08399999991707
172.56099999997332 - RAF callback, animation time: 159.27699999997458
173.70699999992212 - RAF callback, animation time: 160.4219999999259
174.91999999992913 - RAF callback, animation time: 161.62899999994806
176.05700000001434 - RAF callback, animation time: 162.77200000001812
177.18999999999596 - RAF callback, animation time: 163.90599999999722
178.34099999993214 - RAF callback, animation time: 165.05499999993845
179.49299999997947 - RAF callback, animation time: 166.20899999998073
180.62299999996867 - RAF callback, animation time: 167.33899999996993
181.7439999999806 - RAF callback, animation time: 168.45899999998437
182.85800000001018 - RAF callback, animation time: 169.57400000001144
184.0670000000273 - RAF callback, animation time: 170.7809999999199
185.2039999999988 - RAF callback, animation time: 171.9190000000026
186.3979999999401 - RAF callback, animation time: 173.1129999999439
187.5440000000026 - RAF callback, animation time: 174.26000000000386
188.7060000000247 - RAF callback, animation time: 175.42199999991226
189.84199999999873 - RAF callback, animation time: 176.558
191.04299999992236 - RAF callback, animation time: 177.75799999992614
192.1750000000202 - RAF callback, animation time: 178.8899999999103
193.31999999997151 - RAF callback, animation time: 180.03399999997782
194.52100000000883 - RAF callback, animation time: 181.23400000001766
195.68099999992228 - RAF callback, animation time: 182.39599999992606
196.83699999995952 - RAF callback, animation time: 183.5519999999633
197.9749999999285 - RAF callback, animation time: 184.6899999999323
199.16200000000117 - RAF callback, animation time: 185.87700000000495
200.29199999999037 - RAF callback, animation time: 187.00699999999415
201.43999999993412 - RAF callback, animation time: 188.15599999993537
202.5859999999966 - RAF callback, animation time: 189.30199999999786
203.71099999999842 - RAF callback, animation time: 190.4260000000022
204.8579999999447 - RAF callback, animation time: 191.57399999994595
206.01699999997436 - RAF callback, animation time: 192.73299999997562
207.1489999999585 - RAF callback, animation time: 193.8629999999648
208.25999999999567 - RAF callback, animation time: 194.97499999999945
209.39799999996467 - RAF callback, animation time: 196.11299999996845
210.51999999997406 - RAF callback, animation time: 197.23499999997784
211.65499999995063 - RAF callback, animation time: 198.36999999995442
212.80100000001312 - RAF callback, animation time: 199.5160000000169
213.96600000002763 - RAF callback, animation time: 200.68099999991773
213.9799999999923 - Running checks for time: 200, delay: 0.6809999999177307
214.04899999993177 - Checking expectation: [null,0.2,"test","left",100,10]
214.43799999997282 - Ending test
